Initial commit for Dynamic MultiPath Routing project. It includes *almost* all parts for the first phase (per flow balancing).
Two examples are provided.

For simplicity it uses OSPF as a base and its support for equal-cost
multi path.

InterfaceEntry.cc/h
Necessary extension of DmprInterfaceData.

OspfRoutingTableEntry.cc/h
This class is extended of two methods that implemnts the dynamic load
balancing. Each time a RoutingEntry is asked for nextHop, the least
utilised one is returned.

These are the only INET source codes needed to changed.
The rest is implemented in new plug-in modules.


Dmpr.cc/h
Keep track of the current forwarding via DmprForwardingTable.

DmprForwardingTable.cc/h
Stores the current forwarding mapping <5-tuple> -> nextHopInterface.
The "nextHopInterface" contains IP address of the next hop and local
outgoing interface.


RedMarker.cc/.h/.ned
It is based on RedDropper, but instead of dropping it marks IPv4 packets
with ECN. Unlike RedDropper it is not universal queue module, but it is
designed to work IPv4 packets.


The examples are continuously updated to test the implementation status.

examples/dmpr/dynamictest
It is based on OSPF dynamic test.
Only used to test routing traffic to the same destination via different
paths.

examples/dmpr/fulltest

In development. Based on Peyman's scenario for ToN. Meant to test
marking.


examples/dmpr/step1
This on is the most up to date example.
Includes:
 'Dynamic' load balancing for new flows.
  ECN Marking via RedMarker.
  Network route highlighting.



TODO:
Instead of the using the temporary 1flow = increment price for
interface, use Peyman's counting method on the finally working ECN
marking.