DcTcp: Adding alternative way how to calculate marked data
Bytes - works better with unequal segment size but not with packet
reordering and 'delayed' Acks. The 'delayed' Ack (presumably because the
DATA packet was stuck in queue and marked) acknowledges all outstanding
data as "marked".

Number - assumes equal segment size - counts the number of acks as each
should represent the same number of bytes.