DMPR: New balancing mechanism
Becase only DMPR routers participate in Strict Source Routing, the last
nextHop address is ommited.

DmprInterfaceData:
New variables for the new mechanism.

OspfRoutingTableEntry

Utilizes the new variables.

The nextHop is chosen based on the least difference between the fraction
of  'availableLoad' ( 1-congestionFactor) of the sum of all available
loads and the actual forwarding fraction.